+++ jkorm/src/main/kotlin/com/tang/jkorm/utils/id/IdStrategy.kt
@@ -1,0 +1,12 @@
+package com.tang.jkorm.utils.id
+
+import java.lang.reflect.Field
+
+/**
+ * @author Tang
+ */
+interface IdStrategy {
+
+    fun getId(idField: Field): Any
+
+}