+++ jkorm/src/main/kotlin/com/tang/jkorm/utils/id/Ids.kt
@@ -1,0 +1,14 @@
+package com.tang.jkorm.utils.id
+
+/**
+ * @author Tang
+ */
+object Ids {
+
+    private val snowflake = Snowflake(1, 1)
+
+    fun snowflake(): Long {
+        return snowflake.nextId()
+    }
+
+}