+++ jkorm/src/test/kotlin/com/tang/jkorm/session/mapper/RoleMapper.kt
@@ -1,0 +1,9 @@
+package com.tang.jkorm.session.mapper
+
+import com.tang.jkorm.mapper.BaseMapper
+import com.tang.jkorm.session.entity.Role
+
+/**
+ * @author Tang
+ */
+interface RoleMapper : BaseMapper<Role>